SOC-7342: vkil_context_essential is reduced to 16 bytes
and the init API is changed (an handle is always passed)
if invalid, means the handle is not existing and
the field carry some other code.

Notice the handle is provided by the card, and is expected to be the context
address (in WORD unit) on the card. All value below VK_START_VALID_HANDLE
(that is low address not on the sdram) are expected to be invalid, and will be
read as command.

Bug: SOC-7342
Change-Id: I920d1a073eded394c4174a0f94364835deadbeef
Reviewed-on: http://gerrit-ccxsw.broadcom.net/62670
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>
Tested-by: Patrick Rault <patrick.rault@broadcom.com>