SOC-7373: restructure and populate the transfer_buffer function
send_buffer and receive_buffer are depreciated functions.
(testing of them is removed)
all buffer management commands pass to a single transfer function.
The command specify the direction (and if memory upload/downlaod is
required).

To address specific buffer needs, vkil buffer (exposed to the front API)
are separated structures from the internal vk buffer
(thought at this time, the dfinition is the same)

Change-Id: I38ca380a1a7f268c63faca822f4eae0a12e1b48c
Reviewed-on: http://gerrit-ccxsw.broadcom.net/67501
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>