SOC-7373: transfer_buffer function return error code (instead of zero)
On return, especially in call back mode, (that is reading the driver
return queue), we need to know the status of return (e.g. to know if we
need to callback again the function, on (-EAGAIN)...).

The transfer_buffer also is able to take the complete list of command
(instead of only VK_CMD_RUN), thought the flags (BLOCKING or CB) are not
sent to the driver

Change-Id: Ibb071aa1739d23212a8d0b19b27888b4ca5a8bf7
Reviewed-on: http://gerrit-ccxsw.broadcom.net/67814
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>