SOC-7373: add message_id management
Each message is tagged with a 12 bits msg_id which is granted to be
unique among intransit message into the hw system.

This step is required before activating tunneling, B frame scenario,
and more generally when message response flow out of the hw in "random"
fashion (message response order different of the command order).

due to this change, have to remove the "const" on the context handle in
the API (the message ID is managed at context level, the ctx_id being
another field granting the uniqueness of message identification into the
intransit system).

Change-Id: I68343c4dae1c5291a30ed7ba02eff5d12c223b14
Reviewed-on: http://gerrit-ccxsw.broadcom.net/67927
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>