SOC-7941: handle the error on msg_id allocation and driver initialzation
At this time, we allow up 256 simultaneous message in transit.
if we fails to allocate a message id, we treat the error as a
full input queue.

Before, the return from function preset_host2vk_msg was not processed, but
this function was not populating th message on msg_id allocation failure,
so resulting on a faulty message sent to the HW.

failure to initialize the driver (or to find a card) results in a
(-ENODEV) error

Change-Id: Iddb767539b704b88dd6f160bdd1c7bbc62153629
Reviewed-on: http://gerrit-ccxsw.broadcom.net/68532
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>