SOC-7995: modify vkil_read to force waiting for specific msg_id
vkil_wait_probe_read was bailing out after having read a single
message, however this message was not necessary the expected one (msg_id)

Now we start to probe the SW linked list, in case of the message is
already there, ...albeit little chance in Blocking mode). if not:
we "flush" the driver queue (keep calling the driver up to get the
requested msg_id if provided) under timeout constraint.

As an ancilliary, odd waiting loops in the init_com and deinit_com have
been removed, and function called from vkil_api is now vkil_read(...)

Change-Id: Id2cd08e26744457bcc02c644ed0f8dab3aba13da
Reviewed-on: http://gerrit-ccxsw.broadcom.net/69142
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>