SOC-7995: create a vkil_msgid_ctx per device and not per context
Before we were creating a vkil_msgid_ctx per context, however
2 messages going to the same context could share the same msg_id.
In principle we could still differntiate them per their context_id,
...except at the context creation where we don't have this info yet
(context_id is set to VK_NEWCTX)

To prevent that, the vkil_msgid_ctx is moved from the vkilctx a structure
per context, or ffmpeg component) to the vkilctx->devctx (a single
structure per device).

As an ancillary, all the message_id functions are moved from vkil_api.c
to vkil_backend.c

Notice that the msg_id field is overwritten by the driver: The msg_id
inside the vksim is different of the one seen in vkil, and restitute the
vkil's message_id on message return.

The vkil also use the msg_id to pair the incoming message with the
outgoing one.
As well it use this information to pair it with "sw user data", such as
the frame presentation time stamp (pts)

Change-Id: I6dcdfb6bed17aab79c26e36816cbe0d4ff4ead1f
Reviewed-on: http://gerrit-ccxsw.broadcom.net/69811
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>