SOC-7942: Use the buffer prefix and add an aggregated buffers struct
Rather than to redefine the prefix (handle, user_data...)  as it is
currently, nest the vkil_buffer structure in the more specific
buffer structure (packet, surface,...).

The aggregated structure (collection of buffers) is added to handle the
return of multipe buffers by a "process_buffer_done" message.

As well, the buffer type enum used by the vkil_buffer structure is 
prefixed VKIL rather than VK, since we are in the vkil library

The vkil_buffer structure is used at the front end of the vkil (interface
with host such as ffmpeg) while the vk_buffer structure is used at the
backend. Thought these definitions are currently the same, we have 2
differents definition to allow more flexibility in the future).

Change-Id: Id61a1bf3748ff6380b40654a45bf8f6c46b57185
Reviewed-on: http://gerrit-ccxsw.broadcom.net/69957
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>