SOC-8146: set number of plane in transfer command
On buffer transfer function (VK_FID_TRANS_BUF), the command field
(that is host2vk:args[0] field encode the transfer direction (legacy code)
but also the number of "planes" to transfer.

This change use the new vk_buffer_surface structure
(commit a16afc0cfe8cd59d4bc316ef783fa8e667b0a02b).

In the case of the vk_buffer-surface, we set the exact number of plane to
transmit (1, 2 or 4), and since the list seen by the driver
(for sgl list conversion) will be then the bottom of the structure list,
we populate the vk_buffer_surface:planes list in reverse order

(usual order is preserved for the front end api)

also introduce a vkil_size structure to encode the surface size.

Change-Id: Ic2757614d4c030267367e90fd7afd9343ce320e4
Reviewed-on: http://gerrit-ccxsw.broadcom.net/73115
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>