SOC-7942: fix dangling pointer issues in transferring vk_surface
A surface can have up to 4 planes, but we ususally use less (typically
one for YOL2 format to 2 for NV12 format, for progressive mode -
to additional plane are provided to handle the interlace case)

However, not set plane fields, were messing up the driver, with
eventually catastrophic effect (due to some passed invalid address and
size). this in addition of possibly write failure (driver check
consistency address and size consistency since commit
32469b394399451efc47c36f61481c056e30c193)

We force not used planes' field to zero.

Change-Id: Idff6574b18aed8e007b427dfaf45e5b382a19f76
Reviewed-on: http://gerrit-ccxsw.broadcom.net/74732
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>