SOC-8146: handle user_data field at vkil_level whenever existing
When a vkil_buffer is passed thru the vkil front end api, the user_data
is associated to the message id (vkil internal field) send to the card.
that allows to retrieve the user_data associated to the buffer returned
by the response message (without requiring an extra pcie transaction).

The user data filed is also increased to 64 bits
(to allow to store a x86-64 pointer).

Change-Id: If7c9c0f9854428ffd47abf5f44550a22cf5b7ffd
Reviewed-on: http://gerrit-ccxsw.broadcom.net/76530
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Shreesha Rajashekar <shreesha.rajashekar@broadcom.com>
Reviewed-by: Chintan Patel <chintan.patel@broadcom.com>