SOC-8701: Set all possible handles  when output to an AG_BUFFERS
A NULL Handle was meaning a handle list termination, however in the SSIM
case, a message can  have a NULL video output (EOS) and still have a
valid ssim buffer handle (ssim buffer lag by one frame the video output)

So now, we populate a AG_BUFFERS with as many handle as amessage can 
hold (it becomes the vksim responsability to set non valid handle to a
NULL value).

There is no need for the host to set a buffer container (vkil_buffer)
for null handle.

Change-Id: Ib32a220ec4829e2014bfaf65c3eae1d1d18a2c38
Signed-off-by: Shreesha Rajashekar <shreesha.rajashekar@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/77337
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>