SOC-8146: allow device initialization without context and add a unit test
Allow to init the device, without explicit context
creation. That can be useful for ancilliaries device connection, such as
diagnostic tools.

The vkil unit test has never been used so far (and was not working). In
the context of this ticket, it is repaired and a basic device temperature
probing test is added.

Change-Id: I23cd9258d55a1dc948ea3c4f1b73b8eae2db86e2
Reviewed-on: http://gerrit-ccxsw.broadcom.net/80162
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>