SOC-8146: Extend the support for flash util data structure in VKIL
The vkil_set_parameter API obtains the size of the data passed to
it by checking the command. Currently the support is available for
port, video encoder config, video scalar config only. For the other
commands, the size is assumed to be 32-bit size.

The change is made to extend the support for obtaining the flash
util config data structure into the set parameter API. The change
is made to allow the setting of all flash util config in one
set_parameter call.

The flash util config data structure is defined in the
vk_parameters.h header file. The file is exported from the vksim
build.

Change-Id: Iacd640877f921cfeb0a555f4817f2f93a45d9b47
Reviewed-on: http://gerrit-ccxsw.broadcom.net/81184
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>