SOC-9004: add doxygen documentation
To generate the doc, you need to have doxygen installed,

 $ sudo apt-get install doxygen

then run
 $doxygen doxy/Doxyfile

the doc main entry will be in html/index.html.

the added file doxy/readme.dox is the one populating the main page
which describe the overall vksim architecture, when the header.tx
is the one to generate pdf document using the Broadcom corportate
format (essentially disclaimer, copyright and logo). for the later you
need to have pdflatex installed

Change-Id: I90e62fedb8d74344dc3eb4be1803e952e95df4fe
Reviewed-on: http://gerrit-ccxsw.broadcom.net/81729
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>