SOC-9066: ensure buffers size are 4 bytes aligned, add a used_size field
PAX DMA transfer sizes need to be 4 byte aligned, however, when we transfer
a bitstream packet from host, we can't guarantee this requirement,

The preferred solution is to require the host to always provide a 4 bytes
aligned buffer size, and provide the valid used_size and the added field.

we assert on size when not 4 byte aligned. for surface we assert on
stride when not 4 byte aligned.

the alternative could have been to modified the sg list on the Valkyrie
side:
 * That is adding more clutter
 * and so doing, is masking the fact that card can read more memroy than
   wanted.

Change-Id: Ic4b322bda0afc63c404d49a1b0bb3f37f2a4722d
Reviewed-on: http://gerrit-ccxsw.broadcom.net/82188
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>