SOC-9079 repair the native x86 user space only model
This compilation mode got broken in the recent library dependency swap
(it was depending on vksim to work).

The user space model driver is moved into a separate library
from the vkil; vkdrv; since technically the driver is not part of vkil:
it is an utility provided to replace the kernel driver (it could be in
a different workspace, but since it is only a couple line of code, which
anyway are supposed to work only with vkil, there is no practical value
for that)

The library is compiled per default, but to enable the native x86 user
space only model, one has to configure automake as below

 ./configure --enable-drvmodel

and the whole vkil will be using the native x86 user space only model
instead of the kernel driver. Notice in this case, one will also need
to install vksim (code will dynamically link to the vksim library) what
is a given, since the native x86 user space only model use also the x86
model of the vksim

Change-Id: I9222acb26c8b2b08ec3c2433fbe72575dcda8d66
Reviewed-on: http://gerrit-ccxsw.broadcom.net/83279
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>