SOC-8146: Fix the extra flash write call for 4k aligned data size
Fix the extra flash write call for 4k aligned data size.
The flash util app is attempting to write 0 bytes for the case,
when the data size is aligned with 4096 bytes. The flash write
fails and the driver/host reports error.
A check for the write data is added prior to the flash write call
to prevent extra flash write.

Change-Id: I4ac4f91924f034df4ff412ac0f296e6a4bd0eba3
Reviewed-on: http://gerrit-ccxsw.broadcom.net/95971
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Mohamed Ismail Abdul Packir Mohamed <mohamed-ismail.abdul@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>