SOC-7941: vkil return directly vkil_read return value
To allow the host flexibility on returned message error (msg hw_status
flagged as error) are not processed by the vkil, but are conveyed with
linux error code -EADV (Error advised).

vkil_read was previoulsy returning the size of vk2host message read from
the driver (on the model of read), but this value is not used, and vk2host
messages are not exposed to the host application. So for sake of
simplicity vkil_read return
 * 0 on success
 * -EADV on vk2host messag reporting error
 * other negative error code otherwise

Change-Id: I3532a16b138378c42eb5dad9b28b5e3a547ec21f
Reviewed-on: http://gerrit-ccxsw.broadcom.net/122084
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>