SOC-10870: Offline data to include location of next shot change
Instead of just whether the current frame is a shot change, we now
indicate the number of frames to the next known shot change. This
information will be needed to get the GOP structure truncated at
that shot change correct, and can be used in the future for
improved bitrate cap and similar behaviour as we can forecast
future I frames.

Provision has also been included for signalling a future known EOS
which will also be helpful in the future.

Change-Id: I8c9d2db31d6bcbfde574dff5fa95db044a05b2e7
Depends-On: I5c71a8ad4dafe1251fa00d86e0201fc2a6a246b1
Depends-On: I46f3206e2b7aed1279a44d0f74de8c4c9e210232
Depends-On: I27a83cc4d89665d36eee5e0ab00a0bfcadfb8214
Depends-On: Ic6359d27bbb86b441ede41c57c909a7fa4ea988c
Reviewed-on: http://gerrit-ccxsw.broadcom.net/124020
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Graham Rushton <graham.rushton@broadcom.com>