SOC-11765: Remove vkil_deinit on failure from vkil_init()
If vkil_init() fails, it calls vkil_deinit() immediately to clean up.
This means that the caller can't access the context to retrieve error
details. This change removes the call to vkil_deinit() on failure,
transferring this responsibility to the caller once it has interacted
with the context in any required way.

This commit also demotes a vkil generic error reporting message to
DEBUG level, to prevent it muddying the user-facing logs at the
application level now that there is a more specific mechanism for
reporting errors to the user.

Change-Id: I7e3872df710d1b87927ec327b7dbec5f23ac9042
Depends-On: Ic086448b4c907be38b50ce24134d726896d1112c
Reviewed-on: http://gerrit-ccxsw.broadcom.net/124278
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>