SOC-10168: Add Ability To Set and Get Buffer Headers
Reading/writing vksim buffer headers through the M7 and A72 caches
causes many problems, as the buffers aren't aligned on cache line
boundaries and the hooks to invalidate and flush the caches are
complicated to insert on the M7.  As a result we use the vkipc
layer to avoid such issues.

Change-Id: I4cb7be5357e8f17ff95355b71c23cb769c09f08e
Reviewed-on: http://gerrit-ccxsw.broadcom.net/125878
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Tom Miller <tom.miller@broadcom.com>
Reviewed-by: Colin Bell <colin.bell@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Graham Rushton <graham.rushton@broadcom.com>