SOC-11779: add a vkil buffer type to carry field only (not buffer)
The encoder so far has used the vkil_buffer_meta_data to transfer a field
value (such as quality or qp).
Since, it is a field, the buffer reference count doesn't apply, however a
new buffer type is added accordingly to discrimnate between "real" buffer
(handle point to an address), and field (carried into the vkil_buffer_ag
structure).

Change-Id: I0536e4d180ef3696f22459accc2fd6f0a16995b4
Reviewed-on: http://gerrit-ccxsw.broadcom.net/126088
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>