SOC-11258: add a transfer_buffer variant to get transferred size
The original transfer_buffer api is writing the transferred size, or
requested additional into the buffer handle location what is fairly
counter-intuitive, and prevent the buffer to be reused.

The new api has a dedicated field for size for sake of simplicity
The old api is preserved for the time being.

Change-Id: I1e71880925aaf8a00bc2a2f3007ae021ab368e1f
Reviewed-on: http://gerrit-ccxsw.broadcom.net/127547
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>