SOC-12016: encode size on 28 bits instead of 24
It appears that the size of a 4K surface (withh all alignment
constraint) take well over 23 bits, so that 24 bits signed bit is not large
enough.

VK_SIZE_MASK is increased  to 28 bits, from 24 bits, leaving 4 bits for the
flags, which is currently sufficient since it is only used by VK_BUF_EOS

If more bits are required in the future, then probably, it will be better
to increase the message size for those occurrence.

This change is mainly required since commit 5999c625
(SOC-11779: Only decrement ref count on download when complete)

which cast the size on VK_SIZE_MASK to decide whether a buffer need to be
deref or not: the cast size of 4K surface  buffer was then translated into
a negative number.

Change-Id: If3e19eb091526ec1085518c129d0cc7ce97fc7c0
Reviewed-on: http://gerrit-ccxsw.broadcom.net/129601
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>