SOC-12201: remove a possible race condition involved by the use of strerror
The strerror function returns a pointer to a string, the contents of which
are locale-specific. (subclause 7.24.6.2, paragraph 4 [ISO/IEC 9899:2011]).

Thought on the surface the vkil is not multithreaded, the api can be called
from concurrent threads which make the string returned by strerr vulnerable
to corruption.

To prevent such cases, the posix strerror_r function is used instead.
Furthermore, since this one imposes the creation of a user defined buffer,
the function is encapsulated into a new primitive vkil_err which is now
called on error.

The change get rid of the coverity issues CON33-C and ENV30-C related to
the use of strerror

Change-Id: Idf4683a6ff232a1001ce44dca4935d307c233156
Reviewed-on: https://gerrit-ccxsw.broadcom.net/134047
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: CCXSW Coverity <ccxswcoverity@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>
Tested-by: Patrick Rault <patrick.rault@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>