SOC-12201: remove CERT-C warning related to use of errno
CERT-C:2012 rule ERR30-C specifies that errno should be set before using it
It is useless in most case, since we check the errno value only on
failed return (it is then granted errno is set by the called function).

however, since the impact on the code is neglictible, we comply with the
rule.

a #ifdef to accommodate the cmodel in the production code is also removed
(the model is modified to support this).

Change-Id: I8067ba5d4b3dbca3f8c053852055f68f1469ad36
Reviewed-on: https://gerrit-ccxsw.broadcom.net/134230
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: CCXSW Coverity <ccxswcoverity@broadcom.com>
Reviewed-by: Patrick Rault <patrick.rault@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Patrick Rault <patrick.rault@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>